Ensure add/delete operations work for ecs component buffers - Added IsBufferZero() function to Operations.h - Added IsSingleBitSet() function to BitFied.h - Ensured that basic Add/Delete operations of components work for the Modules/ECSTest/ComponentArrays - Other smaller refactors throughout Modules/ECSTest

+++ Engine/Modules/ECSTest/CommonTest.h
@@ -15,7 +15,7 @@ namespace Kargono::ECS
 	constexpr EntityID k_InvalidEntityID{ std::numeric_limits<EntityID>::max() };
 
 	// Component types
-	using ComponentMask = uint8_t;
+	using ComponentMask = uint32_t;
 	constexpr ComponentMask k_MaxComponents{ 32 }; // Arbitrary
 	using ComponentIndex = size_t;
 	using ComponentCount = ComponentIndex;
